cmd/agent-py/requirements.txt
praw
prometheus_client
python_dotenv
pika
google-cloud-storage